share logic in a thunk
